+++ src/api/admin/admin-check-question-answer.ts
@@ -1,5 +1,5 @@
-import { server_axiosInstance } from '../utils/axios';
-import { getCookie } from '../utils/cookies';
+import { server_axiosInstance } from '../../utils/axios';
+import { getCookie } from '../../utils/cookies';
 
 interface AdminCheckQuestionAnswerProps {
   questionId: number;
@@ -9,7 +9,7 @@ export async function AdminCheckQuestionAnswer({ questionId }: AdminCheckQuestio
     questionId,
   };
   try {
-    const response = await server_axiosInstance.post('/api/admin/questions/check', JSON.stringify(data), {
+    const response = await server_axiosInstance.put('/api/admin/questions/check', JSON.stringify(data), {
       headers: {
         Authorization: `Bearer ${getCookie('accessToken')}`,
         'Content-Type': 'application/json',
