+++ src/api/admin/admin-llm-file.ts
@@ -1,4 +1,4 @@
-import { llm_axiosInstance } from '../utils/axios';
+import { llm_axiosInstance } from '../../utils/axios';
 
 interface AdminLLMFileProps {
   type?: string;
