+++ src/ui/pages/admin/login.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Spin } from 'antd';
 import { LoadingOutlined } from '@ant-design/icons';
 import { useNavigate } from 'react-router-dom';
-import { AdminLogin } from '../../../api/admin-login';
+import { AdminLogin } from '../../../api/admin/admin-login';
 
 export default function Login() {
   const [email, setEmail] = useState<string>('');
