+++ src/ui/pages/admin/question-check.tsx
@@ -1,10 +1,10 @@
 import { Divider, Select, Table, TableProps, Tag, AutoComplete, Input } from 'antd';
 import React, { useEffect, useState } from 'react';
-import { adminQuestionCheck } from '../../../api/admin-question-check';
+import { adminQuestionCheck } from '../../../api/admin/admin-question-check';
 import EditModal from '../../components/admin/modal/edit-modal';
 import useCheckQuestionAnswerStore, { QuestionAnswerState } from '../../../store/admin/check-question-answer-store';
 import { searchAutoComplete } from '../../../api/search-auto-complete';
-import { SearchById } from '../../../api/admin-search-by-id';
+import { SearchById } from '../../../api/admin/admin-search-by-id';
 
 const columns: TableProps<QuestionAnswerState>['columns'] = [
   {
